fix: refactored formmatedDate to a util function
